OK, Linux dragging update:
The WindowDraggingUtils.jsm file uses a special function to drag windows
on linux. Apparently the event has to be handed off to the native window
manager while the mouse button is pressed.

This doesn't happen with our minvid window.

Why? I'm not sure. The function we need, `beginWindowMove`, is provided
by the `nsIDOMChromeWindow` interface.

Note that doing `this._window instanceof Ci.nsIDOMChromeWindow` returns
true, so the interface _should_ be applied.

However, the other functions provided by that interface also are not
available on the window that I have.

See here for other stuff in the IDL:
https://dxr.mozilla.org/mozilla-central/source/dom/interfaces/base/nsIDOMChromeWindow.idl

-----

What could this mean? Maybe we just figure out how to position it like
the panel was positioned: bottom-left of the opening window. (but what
happens when that window is destroyed or minimized?)

Let's see if it can be done, first.