Current status: retrying the gtk method w/ostypes fixes applied
See https://github.com/6a68/min-vid/pull/1 for background.

I've cherry-picked the commits from that pull request into this branch,
which is the window-player branch with all commits discarded back to
October 19, when I discovered that the gtk code caused linux to hang.

If this fix works, then we'll have dragging and a topmost window,
without having to worry about WindowDraggingUtils or dragging via a
titlebar shown on hover.

Fingers crossed.